[bugfix] Patrick Bosek patch
This is a simple test to correct an issue when two nodes of the same node-name are siblings and the first is deleted. What happens is the first element doesn't get it's start node deleted and the second does. So  the second element basically ends up with the first one's start element. Which causes problems for the second element's attributes.

svn path=/trunk/eXist/; revision=15541